Properly close colors when logging
This was done in the previous commit. This is a text-only
commit detailing that this was done.

We should always immediately close colors when logging via
`Fore.RESET`, and avoid spaces being colored differently.

Additionally, Fore.RESET is better than using Fore.WHITE, as
it's more standard and will not force white text when the user's
terminal settings are different.

Signed-off-by: Kevin Morris <kevr.gtalk@gmail.com>