Introduce the ability to deduce items into their specific objects
We can now parse an item, which returns an Item type, then call
deduce_specific_object on that Item to convert it into it's
real subclass in O(n) where n = len(NINJA_BASES) time on first call,
and O(1) time on subsequent calls.

On my laptop, first call takes roughly 8.8 milliseconds to run, and
subsequent calls take 0.000001 seconds to run; a fraction of a
millisecond.

Signed-off-by: Kevin Morris <kevr.gtalk@gmail.com>